🎓  Use internal mounted tracker.
Depending on react's internals is a recipe for disaster, and makes this
library wholly dependent on that behaviour.  Fixing this component to
track its mounted state independently should allow for easy interop with
preact/inferno such that this library can be fully utilized in with
something like `preact-compat`.

Also adds further namespacing of an extra `_` to prevent accidental
naming collisions.