💀  Lifecycle Fix
The previous build actually worked fine with ReactDOM, but shallow
rendering doesn't call `componentDidMount` so the tests were failing.
Oops.  Rather than regressing to an compatible internal API, this tracks
the lifecycle differently such that forceUpdate is not called if the
component is not *really* mounted, even though the streams may update
between its attempt to mount and it actually being fully-qualified and
in the DOM.

This makes more sense and stays compatible. :)